Fixing the elaboration of bitwise shift operators
The check on the signedness of the left operand has to use the
**promoted** type. Thanks to @yav for finding the issue.

+++ tests/ci/0108-shifts.c
@@ -8,20 +8,17 @@ int main(void)
   signed char    s1 = 0x7f;
   signed short   s2 = 0x7fff;
 
-  // DEFINED: u1 is unsigned so the result is wrapped around.
-  // NOTE that the result type (promotion of the type of u1)
-  // is signed int, but this is irrelevant.
-  sink = u1 << 25;
+  sink = (unsigned int)u1 << 25;
   // DEFINED: the result is representable in signed int (but << 25 would be UB)
   sink = s1 << 24;
   
   // likewise all of these are DEFINED
-  sink = u2 << 17;
+  sink = (unsigned int)u2 << 17;
   sink = s2 << 16;
   sink = u3 << 1;
 
-  sink = u1 << 31;
-  sink = u2 << 31;
+  sink = (unsigned int)u1 << 31;
+  sink = (unsigned int)u2 << 31;
   sink = u3 << 31;
 
   // see 0333-shift_non_representable.undef.c for the undefined cases
